Sample22/src/com/internousdev/zen/dto/InquiryDTO.java
package com.internousdev.zen.dto;

public class InquiryDTO {

	private String name;
	private String mailaddress;
	private String qtype;
	private String body;

	public String getName() {
		return name;
	}
	public void setName(String name) {
		this.name = name;
	}
	public String getMailaddress() {
		return mailaddress;
	}
	public void setMailaddress(String mailaddress) {
		this.mailaddress = mailaddress;
	}
	public String getQtype() {
		return qtype;
	}
	public void setQtype(String qtype) {
		this.qtype = qtype;
	}
	public String getBody() {
		return body;
	}
	public void setBody(String body) {
		this.body = body;
	}

}
